refactor: apply Copilot review suggestions
- Remove unnecessary .toInt() from jOOQ offset calls
- Make withTenantDsl suspend function, remove nested runBlocking
- Remove redundant inputSchema from jOOQ config (schemata overrides it)

+++ dvmm/dvmm-infrastructure/src/main/kotlin/de/acci/dvmm/infrastructure/projection/BaseProjectionRepository.kt
@@ -62,7 +62,7 @@ public abstract class BaseProjectionRepository<T : Any>(
 
             val items = orderedQuery
                 .limit(pageRequest.size)
-                .offset(pageRequest.offset.toInt())
+                .offset(pageRequest.offset)
                 .fetch()
                 .map { mapRecord(it) }
